% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ceu}
\alias{ceu}
\title{Example data set}
\format{
A data frame with 85 rows and 3 variables:
\describe{
\item{fleet}{fleet}
\item{sw}{standard week}
\item{av.ceu}{average CEU}
}
}
\usage{
ceu
}
\description{
Example data set
}
\keyword{datasets}
