% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_assign_fishing_area_by_coordinates.R
\name{fs_assign_fishing_area_by_coordinates}
\alias{fs_assign_fishing_area_by_coordinates}
\title{A function to help associate each record to a fishing area.}
\usage{
fs_assign_fishing_area_by_coordinates(df, crs, polygon, nearest = TRUE)
}
\arguments{
\item{df}{A data frame with columns "x" and "y" containing coordinates.}

\item{crs}{The coordinate reference system.}

\item{polygon}{A shapefile (class: SpatVector, geometry:polygons). The first column should be fishing area ids.}

\item{nearest}{For records with coordinates that do not overlap with polygons, do you want to return
the name of the nearest polygon? Logical. Default is TRUE.}
}
\description{
A function to help associate each record to a fishing area.
}
