% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_calc_site_score_simple.R
\name{fs_calc_site_score_simple}
\alias{fs_calc_site_score_simple}
\title{Calculates the fisheriescape site scores after all records have been assigned to a spatial reference grid.}
\usage{
fs_calc_site_score_simple(
  df,
  fishing.area.col = NULL,
  year.col = NULL,
  week.col = NULL,
  grid.col = NULL
)
}
\arguments{
\item{df}{A data frame with columns for fishing area id, year, week, and spatial reference grid cell id.}

\item{fishing.area.col}{The name of the column with fishing area ids.}

\item{year.col}{The name of the column with years.}

\item{week.col}{The name of the column with weeks.}

\item{grid.col}{The name of the column with spatial reference grid cell ids.}
}
\description{
Calculates the fisheriescape site scores after all records have been assigned to a spatial reference grid.
}
