% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_fill_col.R
\name{fs_fill_col}
\alias{fs_fill_col}
\title{Fill NAs (replace NAs) in a column based on group summaries.}
\usage{
fs_fill_col(df, group.cols, update.col, fun)
}
\arguments{
\item{df}{A data frame}

\item{group.cols}{Names of columns to group by.}

\item{update.col}{Names of column that should be summarized.}

\item{fun}{The function for summarizing (e.g., mean, max etc.)}
}
\value{
A data frame
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
This function was deprecated. Please use \code{fs_fill_missing} instead.
}
\keyword{internal}
