% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_fill_missing.R
\name{fs_fill_missing}
\alias{fs_fill_missing}
\title{Fisheriescape path to CEU, fill NAs (replace NAs) after summarizing trips}
\usage{
fs_fill_missing(
  df,
  gear.type = NULL,
  vessel.col = NULL,
  year.col = NULL,
  week.col = NULL,
  also.grp = NULL
)
}
\arguments{
\item{df}{The data frame produced by function \code{fs_summarize_trips}, for which duplicate rows have been removed.}

\item{gear.type}{Must be either 'trap' or 'non-trap'.}

\item{vessel.col}{Names of column of vessel (or licence) ids.}

\item{year.col}{Name of column of years.}

\item{week.col}{Name of column of weeks.}

\item{also.grp}{Names of other columns to group by. Optional. E.g., 'fishery', 'gear'}
}
\value{
A data frame
}
\description{
Replaces NAs with either group means or group maximums. See Table 2 in "Building a Fisheriescape: mapping the threat of marine wildlife entanglement in vertical fishing lines in the Gulf of St. Lawrence, Canada".
}
\examples{
\dontrun{
trips<-fs_summarize_trips(df=ziff,
                      gear.type="trap",
                      group.cols=c('fishery.name','gear.name','trip.id','fleet'),
                      gear.col='gear.amount',
                      hour.col='hours.fished')

# restrict columns and remove duplicate rows
COLS<-c('fishery.name','gear.name','fleet','trip.id',
       "cfv",'year','sw',
       'sum.gear','max.hours')

trips2<-trips[,COLS]
trips2<-distinct(trips2)

trips3<-fs_fill_missing(trips2,
                       gear.type='trap',
                       vessel.col='cfv',
                       year.col='year',
                       week.col='sw',
                       also.grp=c("fishery.name" ,"gear.name"))

}
}
