% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_fill_rows.R
\name{fs_fill_rows}
\alias{fs_fill_rows}
\title{Fill NAs (replace NAs) in duplicate rows. Use with caution (see examples)!!}
\usage{
fs_fill_rows(df, group.cols, fill.cols)
}
\arguments{
\item{df}{A data frame}

\item{group.cols}{Names of columns to group by.}

\item{fill.cols}{Names of columns that should be filled.}
}
\value{
A data frame
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
This function was deprecated. Please use \code{fs_summarize_trips} instead.
}
\examples{
Lines <- "ID Value1 Value2 Value3 Value4 Value5 Value6
1 A B C z z NA
1 A B C y NA z
2 A B C NA x NA
2 A B C x NA NA
3 A B C x NA NA
3 A B C x y z"
(DF <- read.table(text = Lines, header = TRUE, as.is = TRUE))
fs_fill_rows(DF,1,2:ncol(DF))# works. Note columns "Value5" and "Value6" for ID "1"
fs_fill_rows(DF,"ID",2:ncol(DF))# works.  Note columns "Value5" and "Value6" for ID "1"
fs_fill_rows(DF,c(1,5),2:ncol(DF)) #works. Compare to previous examples.
fs_fill_rows(DF,"ID","Value1:Value6")# does nothing
}
\keyword{internal}
