% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_fleet_by_coordinates.R
\name{fs_fleet_by_coordinates}
\alias{fs_fleet_by_coordinates}
\title{Assign missing fleets based on geographic coordinates.}
\usage{
fs_fleet_by_coordinates(df, polygon)
}
\arguments{
\item{df}{A data frame}

\item{polygon}{A shapefile (class: SpatVector, geometry:polygons) with column "fleet"}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
This function was deprecated. Please use \code{fs_summarize_trips} instead.
}
\keyword{internal}
