% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_fleet_by_date.R
\name{fs_fleet_by_date}
\alias{fs_fleet_by_date}
\title{Assign missing fleets based on nearest date within range for same cfv.}
\usage{
fs_fleet_by_date(df, match.type = "exact", num.days = NULL)
}
\arguments{
\item{df}{A data frame containing columns: cfv, year, fleet, dateland in format "\%Y\%m\%d"}

\item{match.type}{Match with dates that are either 'exact', 'same.year', 'previous.year'. Default is 'exact' date. For 'same.year' and 'previous.year', 'num.days' must be specified.}

\item{num.days}{Plus and minus days around date to search.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
This function was deprecated. Please use \code{fs_summarize_trips} instead.
}
\keyword{internal}
