% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_pivot2.R
\name{fs_pivot2}
\alias{fs_pivot2}
\title{Fisheriescape pivot table 2 in path to CEU calculations.}
\usage{
fs_pivot2(pivot1, trap.fishery = "yes", group.cols)
}
\arguments{
\item{pivot1}{The output from function \code{fs_pivot1}}

\item{trap.fishery}{Is it a trap fishery? Default is 'yes'.}

\item{group.cols}{Names of columns to group by.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
This function was deprecated. Please use \code{fs_summarize_trips} instead.
}
\keyword{internal}
