% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_pivot4.R
\name{fs_pivot4}
\alias{fs_pivot4}
\title{Fisheriescape not really a pivot, but should be done after pivot table 3 in path to CEU calculations.}
\usage{
fs_pivot4(
  df1,
  trap.fishery,
  pivot3,
  fleet.col,
  prov.col = NULL,
  area.note.col = NULL
)
}
\arguments{
\item{df1}{The ziff data frame.}

\item{trap.fishery}{Is it a trap fishery? Default is 'yes'.}

\item{pivot3}{The output from fs_pivot3}

\item{fleet.col}{Name of fleet column. Should match between df and pivot3}

\item{prov.col}{Currently NULL}

\item{area.note.col}{Currently NULL}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
This function was deprecated. Please use \code{fs_summarize_trips} instead.
}
\keyword{internal}
