% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_proportion_week_fished.R
\name{fs_proportion_week_fished}
\alias{fs_proportion_week_fished}
\title{Fisheriescape path to CEU, calculate the proportion of each week fished}
\usage{
fs_proportion_week_fished(
  df = NULL,
  fish.area.summary = NULL,
  gear.type = NULL,
  week.col = NULL,
  fishing.area.col = NULL,
  also.grp = NULL
)
}
\arguments{
\item{df}{The original data frame of fishing records that includes columns dateland and ctchdate, each formatted as YYYY-MM-DD}

\item{fish.area.summary}{The data frame resulting from \code{fs_summarize_fishing_areas}}

\item{gear.type}{Must be either 'trap' or non.trap'.}

\item{week.col}{Name of week column. Names must match between df and fish.area.summary}

\item{fishing.area.col}{Name of fishing area column. Names must match between df and fish.area.summary}

\item{also.grp}{Optional additional columns to group by. E.g., fishery, gear. Names must match between df and fish.area.summary}
}
\description{
See Table 2 in "Building a Fisheriescape: mapping the threat of marine wildlife entanglement in vertical fishing lines in the Gulf of St. Lawrence, Canada".
}
