% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_site_score.R
\name{fs_site_score}
\alias{fs_site_score}
\title{Fisheriescape calculate site score}
\usage{
fs_site_score(df, fa.poly, grid = NULL)
}
\arguments{
\item{df}{The for.site.score data frame.}

\item{fa.poly}{The fishing area polygon. CRS should be albers. Must contain a column called 'fleet'.}

\item{grid}{Grid shapefile. If NULL, defaults to 'dfo_hex_coffen_smout_cropped.shp'.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
This function was deprecated. Please use \code{fs_summarize_trips} instead.
}
\keyword{internal}
