% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_summarize_fishing_areas.R
\name{fs_summarize_fishing_areas}
\alias{fs_summarize_fishing_areas}
\title{Fisheriescape path to CEU, summarize fishing areas within each time period}
\usage{
fs_summarize_fishing_areas(
  df,
  gear.type = NULL,
  week.col = NULL,
  year.col = NULL,
  fishing.area.col = NULL,
  also.grp = NULL
)
}
\arguments{
\item{df}{The dataframe after running \code{fs_summarize_vessels}}

\item{gear.type}{Must be either 'trap' or non.trap'.}

\item{week.col}{Name of week column.}

\item{year.col}{Name of year column.}

\item{fishing.area.col}{Name of fishing area column.}

\item{also.grp}{Optional additional columns to group by. E.g., fishery, gear}
}
\description{
See Table 2 in "Building a Fisheriescape: mapping the threat of marine wildlife entanglement in vertical fishing lines in the Gulf of St. Lawrence, Canada".
}
