% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_summarize_trips.R
\name{fs_summarize_trips}
\alias{fs_summarize_trips}
\title{Fisheriescape path to CEU, summarize trips within fishing areas}
\usage{
fs_summarize_trips(
  df,
  gear.type = NULL,
  group.cols = NULL,
  gear.col = NULL,
  hour.col = NULL,
  day.col = NULL
)
}
\arguments{
\item{df}{A data frame}

\item{gear.type}{Must be either 'trap' or 'non.trap'. Required.}

\item{group.cols}{Names of columns of grouping variables. Needs to include the trip id and fishing area columns.}

\item{gear.col}{Name of gear column to summarize. Required.}

\item{hour.col}{Name of hours column to summarize. Required.}

\item{day.col}{Name of day column to summarize. Required if gear.type is non.trap.}
}
\value{
A data frame
}
\description{
See Table 2 in "Building a Fisheriescape: mapping the threat of marine wildlife
entanglement in vertical fishing lines in the Gulf of St. Lawrence, Canada".
}
\examples{
\dontrun{
trips<-fs_summarize_trips(df=ziff,
                          gear.type="trap",
                          group.cols=c('fishery.name','gear.name','trip.id','fleet'),
                          gear.col='gear.amount',
                          hour.col='hours.fished')
}
}
