% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fs.data.site.score}
\alias{fs.data.site.score}
\title{Example site score calculations}
\format{
Data frame
}
\usage{
fs.data.site.score
}
\description{
Example site score calculations
}
\keyword{datasets}
