% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_calc_ceu.R
\name{fs_calc_ceu}
\alias{fs_calc_ceu}
\title{Calculate average CEU}
\usage{
fs_calc_ceu(pivot4 = NULL)
}
\arguments{
\item{pivot4}{Object returned from function \code{fs_pivot4}}
}
\value{
A data frame
}
\description{
Uses set cards and length cards.
}
