% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_fill_col.R
\name{fs_fill_col}
\alias{fs_fill_col}
\title{Fill NAs (replace NAs) in a column based on group summaries.}
\usage{
fs_fill_col(df, group.cols, update.col, fun)
}
\arguments{
\item{df}{A data frame}

\item{group.cols}{Names of columns to group by.}

\item{update.col}{Names of column that should be summarized.}

\item{fun}{The function for summarizing (e.g., mean, max etc.)}
}
\value{
A data frame
}
\description{
Fill NAs (replace NAs) in a column based on group summaries.
}
\examples{
#df<-fs_get_data(years=2016:2017,species.sought=144,gclass=1,gearcode=41,nafo=c('4t','4s'))
#test<-fs_fill_col(df,group.cols=c('year','nafodiv'),update.col='nugear',fun=mean)
#summary(df$nugear)
#summary(test$nugear)
#test.fail<-fs_fill_col(df,group.cols=c('year','nafodiv'),update.col=c('nugear','amtgear'),fun=mean)
}
