% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_fill_rows.R
\name{fs_fill_rows}
\alias{fs_fill_rows}
\title{Fill NAs (replace NAs) in duplicate rows.}
\usage{
fs_fill_rows(df, group.cols, fill.cols)
}
\arguments{
\item{df}{A data frame}

\item{group.cols}{Names of columns to group by.}

\item{fill.cols}{Names of columns that should be filled.}
}
\value{
A data frame
}
\description{
Fill NAs (replace NAs) in duplicate rows.
}
\examples{
Lines <- "ID Value1 Value2 Value3 Value4 Value5 Value6
1 A B C z z NA
1 A B C y NA z
2 A B C NA x NA
2 A B C x NA NA
3 A B C x NA NA"
(DF <- read.table(text = Lines, header = TRUE, as.is = TRUE))
fs_fill_rows(DF,1,2:ncol(DF))
fs_fill_rows(DF,'ID',"Value1:Value6")
}
