% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_fleet_by_coordinates.R
\name{fs_fleet_by_coordinates}
\alias{fs_fleet_by_coordinates}
\title{Assign missing fleets based on geographic coordinates.}
\usage{
fs_fleet_by_coordinates(df, polygon)
}
\arguments{
\item{df}{A data frame}

\item{polygon}{A shapefile (class: SpatVector, geometry:polygons) with column "fleet"}
}
\description{
Assign missing fleets based on geographic coordinates.
}
