% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_get_herring_phone_survey.R
\name{fs_get_herring_phone_survey}
\alias{fs_get_herring_phone_survey}
\title{Fisheriescape get herring phone survey data for herring CEU calculations}
\usage{
fs_get_herring_phone_survey(start.year = 2013, one.rope = TRUE)
}
\arguments{
\item{start.year}{The oldest year you want.}

\item{one.rope}{Logical. Do you want to return the one.rope table? Default is TRUE.}
}
\description{
Fisheriescape get herring phone survey data for herring CEU calculations
}
