% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_pivot1.R
\name{fs_pivot1}
\alias{fs_pivot1}
\title{Fisheriescape pivot table 1 in path to CEU calculations.}
\usage{
fs_pivot1(df, keep.cols, gear.col, hour.col = NULL, day.col = NULL)
}
\arguments{
\item{df}{A data frame}

\item{keep.cols}{Names of columns in df to retain in output, if different than 'trip.id','fleet', 'gear.col','hour.col, or 'day.col'}

\item{gear.col}{Name of gear column to summarize. Required.}

\item{hour.col}{Name of hours column to summarize. Optional.}

\item{day.col}{Name of day column to summarize. Optional.}
}
\description{
Fisheriescape pivot table 1 in path to CEU calculations.
}
