% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_pivot3.R
\name{fs_pivot3}
\alias{fs_pivot3}
\title{Fisheriescape pivot table 3 in path to CEU calculations.}
\usage{
fs_pivot3(pivot2, group.cols, id.col = NULL, trap.fishery)
}
\arguments{
\item{pivot2}{The output from function \code{fs_pivot2}}

\item{group.cols}{Names of columns to group by.}

\item{id.col}{Name of column that is the vessel or licence or individual specifier.}

\item{trap.fishery}{Is it a trap fishery? Default is 'yes'.}
}
\description{
Fisheriescape pivot table 3 in path to CEU calculations.
}
