% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_site_score.R
\name{fs_site_score}
\alias{fs_site_score}
\title{Fisheriescape calculate site score}
\usage{
fs_site_score(df, fa.poly)
}
\arguments{
\item{df}{The for.site.score data frame.}

\item{fa.poly}{The fishing area polygon. CRS should be albers. Must contain a column called 'fleet'.}
}
\description{
Fisheriescape calculate site score
}
